refactor: move slugify to dedicated module
