+++ src/pages/ClubFinances.tsx
@@ -2,7 +2,7 @@ import  { useParams, Link } from 'react-router-dom';
 import { ChevronLeft, ArrowUp, ArrowDown, DollarSign, ShoppingBag, Clipboard } from 'lucide-react';
 import PageHeader from '../components/common/PageHeader';
 import { useDataStore } from '../store/dataStore';
-import { formatCurrency, formatDate, slugify } from '../utils/helpers';
+import { formatCurrency, formatDate } from '../utils/helpers';
 
 const ClubFinances = () => {
   const { clubName } = useParams<{ clubName: string }>();
