+++ src/pages/ClubSquad.tsx
@@ -2,7 +2,7 @@ import  { useParams, Link } from 'react-router-dom';
 import { Shield, ChevronLeft, Users, Database, ArrowDown, ArrowUp } from 'lucide-react';
 import PageHeader from '../components/common/PageHeader';
 import { useDataStore } from '../store/dataStore';
-import { formatCurrency, slugify } from '../utils/helpers';
+import { formatCurrency } from '../utils/helpers';
 import { useState } from 'react';
 
 const ClubSquad = () => {
